Fix bash version check
Previously version check required minor number (y in "x.y.z") to be
lower than 1. When a version 5.0.0 of bash was released this check
raised a false positive claiming that the bash version is not bigger
than 4.1.0.

The fix checks the major number along with the minor number so that the
script works on bash 5.0.0.